requirements.txt
certifi==2020.6.20
cffi
chardet==3.0.4
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
gensim==3.8.3
idna==2.10
joblib==0.16.0
karateclub==1.0.17
kiwisolver==1.2.0
matplotlib==3.3.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy
olefile==0.46
pandas==1.1.3
Pillow
pycparser
PyGSP==0.5.1
pynvml==8.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-louvain==0.14
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
threadpoolctl==2.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2
